conf: add package overview docs, testing
Since conf is its own package, it can provide a generally useful
configuration language not coupled to the editor. This simplifies the
package and is a good exercise in writing portable code. The single
test fails, but that's just surfacing existing errors rather than
introducing new ones.

After this, we could move editor-specific code, like the constants
holding a default file path, to either cmd/editor or to a root ollie
package.