spellcheck: conform tests to Go stdlib style
Couple of things here. We don't need to make a separate testing
package as we're not dealing with tricky dependency resolution or
special example use cases. We can maintain tests in the regular
spellcheck package which also makes the code more portable.

Secondly this patch adds context to the error message. Before, our
error message would be

	Wrong distance for hello and cello

This becomes:

        spellcheck_test.go:12: LevDistance("hello", "ccccello") = 4.000000, want 1.000000

I think chapter 11 "Testing" in The Go Programming Language book goes
into this. The idea is to provide the inputs, outputs, and expected
result.